Safari-609.1.11.1/LayoutTests/inspector/audit/basic-boolean-expected.txt

Testing the basic functionality of audits.


== Running test suite: Audit.Basic
-- Running test case: Audit.Basic.Boolean.True
Testing value `true`...
PASS: Result should be "pass".

-- Running test case: Audit.Basic.Boolean.False
Testing value `false`...
PASS: Result should be "fail".
